+++ apps/miranum-forms-webview/src/types.d.ts
@@ -1,3 +1,3 @@
 declare module "@koumoul/vjsf";
-declare module "@koumoul/vjsf/lib/VJsf.js"
+declare module "@koumoul/vjsf/lib/VJsf.js";
 declare module "path";