+++ apps/miranum-extension-pack/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* **\[Miranum Modeler\]** Configuration to enable the alignment of the diagram at the origin (#395)
+
+### Changed
+
+* Update dependencies
+
 ## [0.5.4] - 2023-09-19
 
 ### Changed